@@ -15,8 +15,8 @@
     "lint": "eslint \"**/*.{ts,tsx}\""
   },
   "dependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-scripts": "5.0.1",
     "styled-components": "5.3.3",
     "typescript": "4.6.2",
@@ -34,14 +34,14 @@
     "@storybook/preset-create-react-app": "4.1.0",
     "@storybook/react": "6.4.19",
     "@storybook/testing-library": "0.0.9",
-    "@testing-library/jest-dom": "5.16.2",
-    "@testing-library/react": "12.1.4",
+    "@testing-library/jest-dom": "6.4.2",
+    "@testing-library/react": "14.2.1",
     "@testing-library/user-event": "13.5.0",
     "@types/dom-screen-wake-lock": "^1.0.3",
-    "@types/jest": "27.4.1",
+    "@types/jest": "29.5.12",
     "@types/node": "16.11.26",
-    "@types/react": "17.0.40",
-    "@types/react-dom": "17.0.13",
+    "@types/react": "18.2.61",
+    "@types/react-dom": "18.2.19",
     "@types/styled-components": "5.1.24",
     "eslint": "8.57.0",
     "eslint-config-airbnb": "19.0.4",
